pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yoshibookmark"
version = "0.1.0"
description = "A URL and bookmark management system with intelligent search"
authors = [{name = "YoshiBookmark Contributors"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["bookmarks", "url-management", "search", "semantic-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "playwright>=1.40.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "httpx",
]

[project.scripts]
yoshibookmark = "yoshibookmark.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/yoshibookmark"
Documentation = "https://github.com/yourusername/yoshibookmark/blob/main/README.md"
Repository = "https://github.com/yourusername/yoshibookmark"
Issues = "https://github.com/yourusername/yoshibookmark/issues"

[tool.setuptools]
packages = ["yoshibookmark"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
yoshibookmark = ["web/templates/*", "web/static/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=yoshibookmark",
    "--cov-report=term-missing",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
